feat(profile): add service interface
