+++ backend/src/main/java/com/alligator/market/backend/provider/profile/service/ProviderProfileService.java
@@ -1,31 +1,17 @@
 package com.alligator.market.backend.provider.profile.service;
 
 import com.alligator.market.backend.provider.profile.entity.ProviderProfileEntity;
-import com.alligator.market.backend.provider.profile.repository.ProviderProfileRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
 import java.util.Collection;
 import java.util.List;
 
-
 /**
- * Сервис для работы с профилями провайдеров рыночных данных.
- * Предоставляет базовые операции для получения и сохранения профилей.
+ * Сервис для работы с профилями провайдеров.
  */
-@Service
-@RequiredArgsConstructor
-public class ProviderProfileService {
+public interface ProviderProfileService {
 
-    private final ProviderProfileRepository repository;
+    /** Вернуть все профили провайдеров */
+    List<ProviderProfileEntity> findAll();
 
-    /** Возвращает все профили провайдеров */
-    public List<ProviderProfileEntity> findAll() {
-        return repository.findAll();
-    }
-
-    /** Сохраняет заданную коллекцию профилей */
-    public void saveAll(Collection<ProviderProfileEntity> entities) {
-        repository.saveAll(entities);
-    }
+    /** Сохранить коллекцию профилей */
+    void saveAll(Collection<ProviderProfileEntity> entities);
 }
